[client] refactoring/improve  signUpModal.js readability
